8 byte pointers and 64 byte nodes (#122)
Allows long searches with high hash:

setoption name Hash value 384000
setoption name Threads value 384
position fen rnbqkbnr/pppppppp/8/8/6P1/8/PPPPPP1P/RNBQKBNR b KQkq - 0 1
go infinite

info depth 32 seldepth 120 score cp 152 time 92254087 nodes 2808703365580 nps 30445300 pv d7d5 g4g5 e7e5 d2d4 e5d4 g1f3 f8b4 c1d2 b4d6 f3d4 g8e7 c2c4 c7c5 d4b5 d6e5 c4d5 d8d5 h1g1 b8c6 d2c3 c6d4 f1g2 d5d7 b1a3 e8g8 e2e3 e7f5 d1d2 f5h4 e1c1 h4g2 g1g2

Also a speedup at high thread counts due to cache line allignment:
128 Thread performance (EPYC 9654 x2):

setoption name Threads value 128
setoption name Hash value 64000
go movetime 10000

master:
info depth 15 seldepth 49 score cp 3 time 10001 nodes 205595399 nps 20555558 pv d2d4 g8f6 c2c4 e7e6 b1c3 f8b4 d1c2 e8g8 e2e4 d7d5 e4e5 f6e4 f1d3 c7c5 g1f3

patch:
info depth 14 seldepth 49 score cp 2 time 10000 nodes 234156326 nps 23413428 pv c2c4 e7e5 g2g3 c7c6 g1f3 e5e4 f3d4 d8b6 d4b3 a7a5 d2d3 g8f6 f1g2 f8b4

**Speedup: 13.9%**

Loses Elo at low hash conditions like CCRL blitz:
LTC Non-Reg:
LLR: -2.94 (-2.94,2.94) <-3.50,0.50>
Total: 7536 W: 1517 L: 1666 D: 4353
Ptnml(0-2): 39, 931, 1971, 794, 33
https://tests.montychess.org/tests/view/68da7b8bf2c8ac2e3f959d46

Bench: 1124628

+++ src/mcts/iteration.rs
@@ -132,10 +132,10 @@ fn pick_action(searcher: &Searcher, ptr: NodePtr, node: &Node) -> usize {
         k += 1;
     }
     let mut limit = k.max(searcher.params.min_policy_actions() as usize);
-    let mut thresh = 1u32 << (searcher.params.visit_threshold_power() as u32);
+    let mut thresh = 1u64 << (searcher.params.visit_threshold_power() as u32);
     while node.visits() >= thresh && limit < node.num_actions() {
         limit += 2;
-        thresh = thresh.checked_shl(1).unwrap_or(u32::MAX);
+        thresh = thresh.checked_shl(1).unwrap_or(u64::MAX);
     }
     limit = limit.min(node.num_actions());
 
@@ -147,7 +147,7 @@ fn pick_action(searcher: &Searcher, ptr: NodePtr, node: &Node) -> usize {
             // virtual loss
             let threads = f64::from(child.threads());
             if threads > 0.0 {
-                let visits = f64::from(child.visits());
+                let visits = child.visits() as f64;
                 let q2 = f64::from(q) * visits
                     / (visits + 1.0 + searcher.params.virtual_loss_weight() * (threads - 1.0));
                 q = q2 as f32;
